zfastapi-backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
pydantic==1.10.8
python-multipart==0.0.6
python-dotenv==1.0.0

# MongoDB
motor==3.1.2

# AWS
boto3==1.26.0

# ML & NLP
numpy==1.24.3
scikit-learn==1.2.2

# Transcription
pytesseract==0.3.10
openai-whisper==20230314
moviepy==1.0.3
SpeechRecognition==3.10.0
PyPDF2==3.0.1
python-docx==0.8.11
docx2txt==0.8
beautifulsoup4==4.12.2
pillow==9.5.0

# HTTP Client
requests==2.31.0